guardar id del usuario y guardar posts con promesas
