+++ src/view-screen/contact.js
@@ -2,6 +2,7 @@
 export const showComentarios = (data) => {
   let viewContact = '';
   data.forEach((comentario) => {
+    const date = new Date(comentario.fecha);
     viewContact += `
     <div class="publicaciones">
       <!-- Avatar -->
@@ -10,7 +11,7 @@ export const showComentarios = (data) => {
       <div class="comment-box">
       <section class="comment-head">
         <h6 class="comment-name by-author">${comentario.nombre}</h6>
-        <span>hace 20 minutos</span>
+        <span>${date.getMonth()}</span>
         <p>eliminar</p>
         <i class="fa fa-reply"></i>
         <i class="fa fa-heart"></i>
